did lots with money. new class and no longer a float (now int)

+++ Oregon Trail/src/Equipment.java
@@ -12,7 +12,7 @@ public class Equipment {
 
 	private String name;
 	private int weight; 
-	private float quantity;
+	private int quantity;
 
 	/**
 	 * Creates an Item object containing a name and weight
@@ -20,7 +20,7 @@ public class Equipment {
 	 * @param weight - the weight of the item
 	 * @param quantity - quantity of the item
 	 */
-	public Equipment(String name, int weight, float quantity) {
+	public Equipment(String name, int weight, int quantity) {
 		this.name = name;
 		this.weight = weight;
 		this.quantity = quantity;
@@ -46,7 +46,7 @@ public class Equipment {
 	 * adds the specified amount to the quantity of the item
 	 * @param quantity - the amount to be added
 	 */
-	public void addQuantity(float quantity) {
+	public void addQuantity(int quantity) {
 		this.quantity += quantity;
 	}
 	
@@ -62,7 +62,7 @@ public class Equipment {
 	 * returns the quantity of the item
 	 * @return quantity - the amount of the item in the wagon
 	 */
-	public float getQuantity() {
+	public int getQuantity() {
 		return quantity;
 	}
 }
